Changing the signature to PhiloGL.hasWebGL() again.
There's a big problem with just testing `window.WebGLRenderingContext`
only, and is that we're just looking to see if the browser exposes a
WebGL API, but we don't check anything about whether WebGL is actually
supported or not. In machines that don't support WebGL, we will still
get true for !!window.WebGLRenderingContext on browsers like Ch9 or FF4,
that expose the WebGL API even if our computer is not capable of
creating WebGL graphics at all. Trying to create a context is a much deeper check than just
checking for the API. I noticed this from the comments the people had
after I did the first change.